Update check-manifest invocation
check-manifest 0.42 updated `--ignore` handling to be more consistent with MANIFEST.in global-exclude directives.  As a result, `tests*` no longer matches files inside a `tests` directory, and you have to use `tests/**` instead.

As an extra complication, check-manifest drops Python 2.7 support.  Since you still have py27 in this tox.ini, this means you get an older check-manifest 0.41, and you have to tell it to ignore the empty `tests` directory that remains after you exclude all the files in it.

For details see the second half of https://github.com/mgedmin/check-manifest/issues/98

I apologize for the inconvenience!